+++ packages/verified-fetch/test/abort-handling.spec.ts
@@ -1,10 +1,13 @@
 import { dagCbor } from '@helia/dag-cbor'
 import { type DNSLinkResolveResult, type IPNS, type IPNSResolveResult } from '@helia/ipns'
-import { type UnixFSStats, unixfs } from '@helia/unixfs'
+import { unixfs } from '@helia/unixfs'
 import { stop, type ComponentLogger, type Logger } from '@libp2p/interface'
 import { prefixLogger, logger as libp2pLogger } from '@libp2p/logger'
 import { createEd25519PeerId } from '@libp2p/peer-id-factory'
 import { expect } from 'aegir/chai'
+import browserReadableStreamToIt from 'browser-readablestream-to-it'
+import { fixedSize } from 'ipfs-unixfs-importer/chunker'
+import drain from 'it-drain'
 import { CID } from 'multiformats/cid'
 import pDefer, { type DeferredPromise } from 'p-defer'
 import Sinon from 'sinon'
@@ -31,11 +34,9 @@ describe.skip('abort-handling', function () {
   /**
    * Stubbed networking components
    */
-  let blockRetriever: StubbedInstance<Required<BlockBroker>>
+  let blockRetriever: StubbedInstance<Required<Pick<BlockBroker, 'retrieve'>>>
   let dnsLinkResolver: Sinon.SinonStub<any[], Promise<DNSLinkResolveResult>>
   let peerIdResolver: Sinon.SinonStub<any[], Promise<IPNSResolveResult>>
-  let unixFsCatStub: Sinon.SinonStub<any[], AsyncIterable<Uint8Array>>
-  let unixFsStatStub: Sinon.SinonStub<any[], Promise<UnixFSStats>>
 
   /**
    * used as promises to pass to makeAbortedRequest that will abort the request as soon as it's resolved.
@@ -43,14 +44,10 @@ describe.skip('abort-handling', function () {
   let blockBrokerRetrieveCalled: DeferredPromise<void>
   let dnsLinkResolverCalled: DeferredPromise<void>
   let peerIdResolverCalled: DeferredPromise<void>
-  let unixFsStatCalled: DeferredPromise<void>
-  let unixFsCatCalled: DeferredPromise<void>
 
   beforeEach(async () => {
     peerIdResolver = sandbox.stub()
     dnsLinkResolver = sandbox.stub()
-    unixFsCatStub = sandbox.stub()
-    unixFsStatStub = sandbox.stub()
     peerIdResolverCalled = pDefer()
     dnsLinkResolverCalled = pDefer()
     blockBrokerRetrieveCalled = pDefer()
@@ -63,7 +60,7 @@ describe.skip('abort-handling', function () {
       peerIdResolverCalled.resolve()
       return getAbortablePromise(options.signal)
     })
-    blockRetriever = stubInterface<Required<BlockBroker>>({
+    blockRetriever = stubInterface<Required<Pick<BlockBroker, 'retrieve'>>>({
       retrieve: sandbox.stub().callsFake(async (cid, options) => {
         blockBrokerRetrieveCalled.resolve()
         return getAbortablePromise(options.signal)
@@ -128,39 +125,137 @@ describe.skip('abort-handling', function () {
     expect(blockRetriever.retrieve.callCount).to.equal(1)
   })
 
-  it('should abort a request during unixfs.stat call', async function () {
+  it('should abort a request while loading a file root', async function () {
     const fs = unixfs(helia)
-    const fileCid = await fs.addBytes(Uint8Array.from([0, 1, 2, 3]))
+
+    // add a file with a very small chunk size - this is to ensure we end up
+    // with a DAG that contains a root and some leaf nodes
+    const fileCid = await fs.addBytes(Uint8Array.from([0, 1, 2, 3]), {
+      chunker: fixedSize({ chunkSize: 2 })
+    })
     const directoryCid = await fs.addDirectory()
     const cid = await fs.cp(fileCid, directoryCid, 'index.html')
 
-    await expect(makeAbortedRequest(verifiedFetch, [cid], unixFsStatCalled.promise)).to.eventually.be.rejectedWith('aborted')
+    const leaf1 = CID.parse('bafkreifucp2h2e7of7tmqrns5ykbv6a55bmn6twfjgsyw6lqxolgiw6i2i')
+    const leaf2 = CID.parse('bafkreihosbapmxbudbk6a4h7iohlb2u5lobrwkrme4h3p32zfv2qichdwm')
 
-    expect(peerIdResolver.callCount).to.equal(0) // not called because parseResource never passes the resource to parseUrlString
-    expect(dnsLinkResolver.callCount).to.equal(0) // not called because parseResource never passes the resource to parseUrlString
-    expect(blockRetriever.retrieve.callCount).to.equal(0) // not called because the blockstore has the content
-    expect(unixFsStatStub.callCount).to.equal(1)
-    expect(unixFsCatStub.callCount).to.equal(0) // not called because the request was aborted during .stat call
-  })
+    // file root
+    await expect(helia.blockstore.has(fileCid))
+      .to.eventually.be.true()
 
-  it('should abort a request during unixfs.cat call', async function () {
-    const fs = unixfs(helia)
-    const fileCid = await fs.addBytes(Uint8Array.from([0, 1, 2, 3]))
-    const directoryCid = await fs.addDirectory()
-    const cid = await fs.cp(fileCid, directoryCid, 'index.html')
+    // leaf nodes
+    await expect(helia.blockstore.has(leaf1))
+      .to.eventually.be.true()
+    await expect(helia.blockstore.has(leaf2))
+      .to.eventually.be.true()
 
-    // override the default fake set in beforeEach that would timeout.
-    unixFsStatStub.callsFake(async (cid, options) => {
-      unixFsStatCalled.resolve()
-      return fs.stat(cid, options)
+    const fileRootGot = pDefer()
+    const blockstoreGetSpy = Sinon.stub(helia.blockstore, 'get')
+    blockstoreGetSpy.callsFake(async (cid, options) => {
+      if (cid.equals(fileCid)) {
+        fileRootGot.resolve()
+      }
+
+      return blockstoreGetSpy.wrappedMethod.call(helia.blockstore, cid, options)
     })
 
-    await expect(makeAbortedRequest(verifiedFetch, [cid], unixFsCatCalled.promise)).to.eventually.be.rejectedWith('aborted')
+    await expect(makeAbortedRequest(verifiedFetch, [cid], fileRootGot.promise))
+      .to.eventually.be.rejectedWith('aborted')
 
-    expect(peerIdResolver.callCount).to.equal(0) // not called because parseResource never passes the resource to parseUrlString
-    expect(dnsLinkResolver.callCount).to.equal(0) // not called because parseResource never passes the resource to parseUrlString
-    expect(blockRetriever.retrieve.callCount).to.equal(0) // not called because the blockstore has the content
-    expect(unixFsStatStub.callCount).to.equal(1)
-    expect(unixFsCatStub.callCount).to.equal(1)
+    // not called because parseResource never passes the resource to
+    // parseUrlString
+    expect(peerIdResolver.callCount).to.equal(0)
+
+    // not called because parseResource never passes the resource to
+    // parseUrlString
+    expect(dnsLinkResolver.callCount).to.equal(0)
+
+    // not called because the blockstore has the content
+    expect(blockRetriever.retrieve.callCount).to.equal(0)
+
+    // the file root was loaded
+    expect(blockstoreGetSpy.getCalls().map(call => call.args[0].toString()))
+      .to.include(fileCid.toString())
+
+    // the leaf nodes were not loaded because the request was aborted
+    // after the root node was loaded
+    expect(blockstoreGetSpy.getCalls().map(call => call.args[0].toString()))
+      .to.not.include(leaf1.toString())
+    expect(blockstoreGetSpy.getCalls().map(call => call.args[0].toString()))
+      .to.not.include(leaf2.toString())
+  })
+
+  it('should abort a request while loading file data', async function () {
+    const fs = unixfs(helia)
+
+    // add a file with a very small chunk size - this is to ensure we end up
+    // with a DAG that contains a root and some leaf nodes
+    const fileCid = await fs.addBytes(Uint8Array.from([0, 1, 2, 3]), {
+      chunker: fixedSize({ chunkSize: 2 })
+    })
+    const directoryCid = await fs.addDirectory()
+    const cid = await fs.cp(fileCid, directoryCid, 'index.html')
+
+    const leaf1 = CID.parse('bafkreifucp2h2e7of7tmqrns5ykbv6a55bmn6twfjgsyw6lqxolgiw6i2i')
+    const leaf2 = CID.parse('bafkreihosbapmxbudbk6a4h7iohlb2u5lobrwkrme4h3p32zfv2qichdwm')
+
+    // file root
+    await expect(helia.blockstore.has(fileCid))
+      .to.eventually.be.true()
+
+    // leaf nodes
+    await expect(helia.blockstore.has(leaf1))
+      .to.eventually.be.true()
+    await expect(helia.blockstore.has(leaf2))
+      .to.eventually.be.true()
+
+    const leaf1Got = pDefer()
+    let leaf2Loaded = false
+    const blockstoreGetSpy = Sinon.stub(helia.blockstore, 'get')
+    blockstoreGetSpy.callsFake(async (cid, options) => {
+      if (cid.equals(leaf1)) {
+        leaf1Got.resolve()
+      }
+
+      const b = await blockstoreGetSpy.wrappedMethod.call(helia.blockstore, cid, options)
+
+      if (cid.equals(leaf2)) {
+        leaf2Loaded = true
+      }
+
+      return b
+    })
+
+    const response = await makeAbortedRequest(verifiedFetch, [cid], leaf1Got.promise)
+
+    if (response.body == null) {
+      throw new Error('Body was not set')
+    }
+
+    // error occurs during streaming response
+    await expect(drain(browserReadableStreamToIt(response.body)))
+      .to.eventually.be.rejectedWith('aborted')
+
+    // not called because parseResource never passes the resource to
+    // parseUrlString
+    expect(peerIdResolver.callCount).to.equal(0)
+
+    // not called because parseResource never passes the resource to
+    // parseUrlString
+    expect(dnsLinkResolver.callCount).to.equal(0)
+
+    // not called because the blockstore has the content
+    expect(blockRetriever.retrieve.callCount).to.equal(0)
+
+    // the file root was loaded
+    expect(blockstoreGetSpy.getCalls().map(call => call.args[0].toString()))
+      .to.include(fileCid.toString())
+
+    // the first leaf was loaded
+    expect(blockstoreGetSpy.getCalls().map(call => call.args[0].toString()))
+      .to.include(leaf1.toString())
+
+    // the signal was aborted before the second leaf was loaded
+    expect(leaf2Loaded).to.be.false()
   })
 })